refactor: move the filtering logic to Facility Page root
